final tochup

+++ ProductManagementSystemAPI/ApiBased/ProductManagementSystem/Controllers/CategoryController.cs
@@ -14,21 +14,43 @@ namespace ProductManagementSystem.Controllers
         {
             this.service = service;
         }
+
+        bool IsAdmin()
+        {
+            return Request.Cookies["admin"] == "true";
+        }
+        bool IsCustomer()
+        {
+            return Request.Cookies["customer"] == "true";
+        }
+
         [HttpGet("all")]
         public IActionResult GetAll()
         {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login");
+            }
             var data = service.GetAll();
             return Ok(data);
         }
         [HttpGet("{id}")]
         public IActionResult Find(int id)
         {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login");
+            }
             var data = service.Find(id);
             return Ok(data);
         }
         [HttpPost("add")]
         public IActionResult Add(CategoryDTO c)
         {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login As Admin");
+            }
             var res = service.Add(c);
             if (res == true)
             {
@@ -39,10 +61,14 @@ namespace ProductManagementSystem.Controllers
                 return BadRequest(res);
             }
         }
-        [HttpPost("update")]
-        public IActionResult Update(CategoryDTO c)
+        [HttpPost("update/{id}")]
+        public IActionResult Update(int id,CategoryDTO c)
         {
-            var res = service.Update(c);
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login As Admin");
+            }
+            var res = service.Update(id,c);
             if (res == true)
             {
                 return Ok(res);
@@ -52,9 +78,13 @@ namespace ProductManagementSystem.Controllers
                 return BadRequest(res);
             }
         }
-        [HttpPost("delete/{id}")]
+        [HttpDelete("delete/{id}")]
         public IActionResult Delete(int id)
         {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login As Admin");
+            }
             var res = service.Delete(id);
             if (res == true)
             {
@@ -65,5 +95,63 @@ namespace ProductManagementSystem.Controllers
                 return BadRequest(res);
             }
         }
+
+        [HttpGet("fawp")]
+        public IActionResult FindAllWithProducts()
+        {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login As Admin");
+            }
+            var res=service.FindAllWithProducts();
+            if (res != null)
+            {
+                return Ok(res);
+            }
+            return NotFound("Data not found");
+
+        }
+        [HttpPost("fwn/{name}")]
+        public IActionResult FindByName(string name)
+        {
+            if (!IsAdmin()|| !IsCustomer())
+            {
+                return Unauthorized("Please Login");
+            }
+            var res=service.FindByName(name);
+            if(res != null)
+            {
+                return Ok(res);
+            }
+            return NotFound("Name doesn't contain any category");
+        }
+        [HttpGet("hwp")]
+        public IActionResult HighestProduct() {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login");
+            }
+            var res=service.HighestProdsucts();
+            if (res != null) { 
+            return Ok(res);
+            }
+            return NotFound("Error!");
+
+        }
+        [HttpPost("fwp/{id}")]
+        public IActionResult FindWithProduct(int id)
+        {
+            if (!IsAdmin())
+            {
+                return Unauthorized("Please Login");
+            }
+            var res = service.FindWithProducts(id);
+            if (res != null) 
+            { 
+                return Ok(res); 
+            }
+            return NotFound("id wrong");
+        }
+
     }
 }
